feat(maths): #76 move math to a dedicated module
